Boost.Chrono: Version 0.2.6 * disable VC++ foolishness * Update Jamfiles to manage with dll * removal of conversion warning in test_duration * manage with MSVC reporting a warning instead of an error when there is an integral constant overflow
SVN-Revision: 58324

+++ chrono/libs/chrono/test/test_duration.cpp
@@ -179,7 +179,7 @@ int main()
     //tmpl2(3,3);  // Will not compile, 3 is not implicitly convertible to any `duration`
 
     {
-    double r = milliseconds(3) / milliseconds(3);
+    double r = double(milliseconds(3) / milliseconds(3));
     duration<double, boost::milli> d = milliseconds(3) * 2.5;
     duration<double, boost::milli> d2 = 2.5 * milliseconds(3) ;
     duration<double, boost::milli> d3 = milliseconds(3) / 2.5;
